Move 2 stories for alignment with updated folder structure
